Adicionado busca por CPF na pessoa fisica

+++ src/main/java/br/com/stabulith/bancostabulith/services/PessoaFisicaService.java
@@ -1,16 +1,20 @@
 package br.com.stabulith.bancostabulith.services;
 
+import br.com.stabulith.bancostabulith.entities.Documento;
 import br.com.stabulith.bancostabulith.entities.PessoaFisica;
 import br.com.stabulith.bancostabulith.models.PessoaFisicaDTO;
 import br.com.stabulith.bancostabulith.repositories.PessoaFisicaRepository;
 import br.com.stabulith.bancostabulith.utils.MapperUtil;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.UUID;
 
+@Log4j2
 @Service
 public class PessoaFisicaService {
     @Autowired
@@ -25,16 +29,40 @@ public class PessoaFisicaService {
     }
 
     public PessoaFisicaDTO buscarPorId(UUID id) {
-        return mapperUtil.convertTo(repository.findById(id).get(), PessoaFisicaDTO.class);
+        return mapperUtil.convertTo(repository.findById(id).orElse(new PessoaFisica()), PessoaFisicaDTO.class);
     }
 
     public PessoaFisicaDTO salvar(PessoaFisicaDTO dto) {
-        dto.setDataModificacao(LocalDateTime.now());
-        PessoaFisica save = repository.save(mapperUtil.convertTo(dto, PessoaFisica.class));
-        return mapperUtil.convertTo(save, PessoaFisicaDTO.class);
+        PessoaFisica pessoaFisicaSalvo = new PessoaFisica();
+        try {
+            Optional<UUID> uuidOptional = Optional.ofNullable(dto.getId());
+
+            if (uuidOptional.isPresent()){
+                dto.setDataModificacao(LocalDateTime.now());
+            }
+
+            pessoaFisicaSalvo = repository.save(mapperUtil.convertTo(dto, PessoaFisica.class));
+
+        } catch (Exception e) {
+            log.error(e.getMessage());
+            return null;
+        }
+        return mapperUtil.convertTo(pessoaFisicaSalvo, PessoaFisicaDTO.class);
     }
 
-    public void excluir(UUID id) {
-       repository.deleteById(id);
+    public boolean excluir(UUID id) {
+        try {
+            if (!repository.existsById(id)) {
+                return false;
+            }
+            repository.deleteById(id);
+        } catch (Exception e) {
+            log.error(e.getMessage());
+        }
+        return true;
+    }
+
+    public PessoaFisicaDTO buscarPorCpf(long cpf) {
+        return mapperUtil.convertTo(repository.findPessoaFisicaByDocumentosByCpf(cpf).orElse(new PessoaFisica()), PessoaFisicaDTO.class);
     }
 }
